feat(dataset): do not tokenize actions
leave the actions as continuous because we can easily tokenize them on the fly as needed

+++ src/cogelot/structures/model.py
@@ -141,10 +141,10 @@ class PreprocessedInstance(BaseModel, PydanticHFDatasetMixin):
                     },
                 },
                 "actions": {
-                    "pose0_position": datasets.Array2D(shape=(None, 3), dtype="int64"),
-                    "pose0_rotation": datasets.Array2D(shape=(None, 4), dtype="int64"),
-                    "pose1_position": datasets.Array2D(shape=(None, 3), dtype="int64"),
-                    "pose1_rotation": datasets.Array2D(shape=(None, 4), dtype="int64"),
+                    "pose0_position": datasets.Array2D(shape=(None, 3), dtype="float32"),
+                    "pose0_rotation": datasets.Array2D(shape=(None, 4), dtype="float32"),
+                    "pose1_position": datasets.Array2D(shape=(None, 3), dtype="float32"),
+                    "pose1_rotation": datasets.Array2D(shape=(None, 4), dtype="float32"),
                 },
             }
         )
